requirements-dev.txt
-r requirements.txt
flake8
pytest
pytest-cov
pytest-flask